Refactor mmultiscript, msubsup, munderover, scriptbase into common and chtml components
