+++ mathjax3-ts/output/chtml/Wrappers/scriptbase.ts
@@ -16,27 +16,19 @@
  */
 
 /**
- * @fileoverview  Implements the a base class for CHTMLmsubsup and CHTMLmunderover
+ * @fileoverview  Implements the a base class for CHTMLmsubsup, CHTMLmunderover
  *                and their relatives.  (Since munderover can become msubsup
- *                when movablelimits is set, munderoer needs to be able to
+ *                when movablelimits is set, munderover needs to be able to
  *                do the same thing as msubsup in some cases.)
  *
  * @author dpvc@mathjax.org (Davide Cervone)
  */
 
-import {CHTMLWrapper} from '../Wrapper.js';
-import {CHTMLWrapperFactory} from '../WrapperFactory.js';
-import {CHTMLmo} from './mo.js';
+import {CHTMLWrapper, CHTMLConstructor} from '../Wrapper.js';
+import {CommonScriptbase, CommonScriptbaseInterface} from '../../common/Wrappers/scriptbase.js';
 import {MmlMsubsup} from '../../../core/MmlTree/MmlNodes/msubsup.js';
-import {MmlNode} from '../../../core/MmlTree/MmlNode.js';
 import {BBox} from '../BBox.js';
 import {StyleData, StyleList} from '../../common/CssStyles.js';
-import {DIRECTION} from '../FontData.js';
-
-/*
- * Mutliply italic correction by this much (improve horizontal shift for italic characters)
- */
-const DELTA = 1.5;
 
 /*****************************************************************/
 /**
@@ -47,59 +39,15 @@ const DELTA = 1.5;
  * @template T  The Text node class
  * @template D  The Document class
  */
-export class CHTMLscriptbase<N, T, D> extends CHTMLWrapper<N, T, D> {
+export class CHTMLscriptbase<N, T, D> extends
+CommonScriptbase<N, T, D, CHTMLWrapper<N, T, D>, CHTMLConstructor<N, T, D>>(CHTMLWrapper) {
+
     public static kind = 'scriptbase';
 
     /**
      * Set to true for munderover/munder/mover/msup (Appendix G 13)
      */
     public static useIC: boolean = false;
-
-    /**
-     * The core mi or mo of the base (or the base itself if there isn't one)
-     */
-    protected baseCore: CHTMLWrapper<N, T, D>;
-
-    /**
-     * @return {CHTMLWrapper}  The base element's wrapper
-     */
-    public get baseChild() {
-        return this.childNodes[(this.node as MmlMsubsup).base];
-    }
-
-    /**
-     * @return {CHTMLWrapper}  The script element's wrapper (overridden in subclasses)
-     */
-    public get script() {
-        return this.childNodes[1];
-    }
-
-    /**
-     * @override
-     */
-    constructor(factory: CHTMLWrapperFactory<N, T, D>, node: MmlNode, parent: CHTMLWrapper<N, T, D> = null) {
-        super(factory, node, parent);
-        //
-        //  Find the base core
-        //
-        let core = this.baseCore = this.childNodes[0];
-        if (!core) return;
-        while (core.childNodes.length === 1 &&
-               (core.node.isKind('mrow') || core.node.isKind('TeXAtom') ||
-                core.node.isKind('mstyle') || core.node.isKind('mpadded') ||
-                core.node.isKind('mphantom') || core.node.isKind('semantics'))) {
-            core = core.childNodes[0];
-            if (!core) return;
-        }
-        if (!('noIC' in core)) return;
-        this.baseCore = core;
-        //
-        //  Check if the base is a mi or mo that needs italic correction removed
-        //
-        if (!(this.constructor as typeof CHTMLscriptbase).useIC) {
-            (core as CHTMLmo<N, T, D>).noIC = true;
-        }
-    }
 
     /**
      * This gives the common output for msub and msup.  It is overriden
@@ -119,162 +67,6 @@ export class CHTMLscriptbase<N, T, D> extends CHTMLWrapper<N, T, D> {
     }
 
     /**
-     * This gives the common bbox for msub and msup.  It is overriden
-     * for all the others (msubsup, munder, mover, munderover).
-     *
-     * @override
-     */
-    public computeBBox(bbox: BBox) {
-        const basebox = this.baseChild.getBBox();
-        const scriptbox = this.script.getBBox();
-        const [x, y] = this.getOffset(basebox, scriptbox);
-        bbox.append(basebox);
-        bbox.combine(scriptbox, bbox.w + x, y);
-        bbox.w += this.font.params.scriptspace;
-        bbox.clean();
-    }
-
-    /**
-     * @return {boolean}  True if the base is an mi, mn, or mo (not a largeop) consisting of a single character
-     */
-    protected isCharBase() {
-        let base = this.baseChild;
-        if ((base.node.isKind('mstyle') || base.node.isKind('mrow')) && base.childNodes.length === 1) {
-            base = base.childNodes[0];
-        }
-        return ((base.node.isKind('mo') || base.node.isKind('mi') || base.node.isKind('mn')) &&
-                base.bbox.rscale === 1 && base.getText().length === 1 &&
-                !base.node.attributes.get('largeop'));
-    }
-
-    /***************************************************************************/
-    /*
-     *  Methods for sub-sup nodes
-     */
-
-    /**
-     * Get the shift for the script (implemented in subclasses)
-     *
-     * @param {BBox} bbox   The bounding box of the base element
-     * @param {BBox} sbox   The bounding box of the script element
-     * @return {number[]}   The horizontal and vertical offsets for the script
-     */
-    protected getOffset(bbox: BBox, sbox: BBox) {
-        return [0, 0];
-    }
-
-    /**
-     * Get the shift for a subscript (TeXBook Appendix G 18ab)
-     *
-     * @param {BBox} bbox   The bounding box of the base element
-     * @param {BBox} sbox   The bounding box of the superscript element
-     * @return {number}     The vertical offset for the script
-     */
-    protected getV(bbox: BBox, sbox: BBox) {
-        const tex = this.font.params;
-        const subscriptshift = this.length2em(this.node.attributes.get('subscriptshift'), tex.sub1);
-        return Math.max(
-            this.isCharBase() ? 0 : bbox.d + tex.sub_drop * sbox.rscale,
-            subscriptshift,
-            sbox.h * sbox.rscale - (4/5) * tex.x_height
-        );
-    }
-
-    /**
-     * Get the shift for a superscript (TeXBook Appendix G 18acd)
-     *
-     * @param {BBox} bbox   The bounding box of the base element
-     * @param {BBox} sbox   The bounding box of the superscript element
-     * @return {number}     The vertical offset for the script
-     */
-    protected getU(bbox: BBox, sbox: BBox) {
-        const tex = this.font.params;
-        const attr = this.node.attributes.getList('displaystyle', 'texprimestyle', 'superscriptshift');
-        const p = (attr.displaystyle ? tex.sup1 : attr.texprimestyle ? tex.sup3 : tex.sup2);
-        const superscriptshift = this.length2em(attr.superscriptshift, p);
-        return Math.max(
-            this.isCharBase() ? 0 : bbox.h - tex.sup_drop * sbox.rscale,
-            superscriptshift,
-            sbox.d * sbox.rscale + (1/4) * tex.x_height
-        );
-    }
-
-    /***************************************************************************/
-    /*
-     *  Methods for under-over nodes
-     */
-
-    /**
-     * @return {boolean}  True if the base has movablelimits (needed by munderover)
-     */
-    protected hasMovableLimits() {
-        const display = this.node.attributes.get('displaystyle');
-        return (!display && (this.node.getProperty('movablelimits') ||
-                             this.node.attributes.get('movablelimits') ||
-                             this.baseChild.coreMO().node.attributes.get('movablelimits')));
-    }
-
-    /**
-     * Get the separation and offset for overscripts (TeXBoox Appendix G 13, 13a)
-     *
-     * @param {BBox} basebox  The bounding box of the base
-     * @param {BBox} overbox  The bounding box of the overscript
-     * @return {numner[]}     The separation between their boxes, and the offset of the overscript
-     */
-    protected getOverKU(basebox: BBox, overbox: BBox) {
-        const accent = this.node.attributes.get('accent') as boolean;
-        const tex = this.font.params;
-        const d = overbox.d * overbox.rscale;
-        const k = (accent ? tex.rule_thickness :
-                   Math.max(tex.big_op_spacing1, tex.big_op_spacing3 - Math.max(0, d))) -
-            (this.baseChild.node.isKind('munderover') ? .1 : 0);
-        return [k, basebox.h + k + d];
-    }
-
-    /**
-     * Get the separation and offset for underscripts (TeXBoox Appendix G 13, 13a)
-     *
-     * @param {BBox} basebox   The bounding box of the base
-     * @param {BBox} underbox  The bounding box of the underscript
-     * @return {numner[]}      The separation between their boxes, and the offset of the underscript
-     */
-    protected getUnderKV(basebox: BBox, underbox: BBox) {
-        const accent = this.node.attributes.get('accentunder') as boolean;
-        const tex = this.font.params;
-        const h = underbox.h * underbox.rscale;
-        const k = (accent ? tex.rule_thickness :
-                   Math.max(tex.big_op_spacing2, tex.big_op_spacing4 - h)) -
-            (this.baseChild.node.isKind('munderover') ? .1 : 0);
-        return [k, -(basebox.d + k + h)];
-    }
-
-    /**
-     * @param {BBox[]} boxes    The bounding boxes whose offsets are to be computed
-     * @param {number[]} delta  The initial x offsets of the boxes
-     * @return {number[]}       The actual offsets needed to center the boxes in the stack
-     */
-    protected getDeltaW(boxes: BBox[], delta: number[] = [0, 0, 0]) {
-        const align = this.node.attributes.get('align');
-        const widths = boxes.map(box => box.w * box.rscale);
-        const w = Math.max(...widths);
-        const dw = [];
-        let m = 0;
-        for (const i of widths.keys()) {
-            dw[i] = (align === 'center' ? (w - widths[i]) / 2 :
-                     align === 'right' ? w - widths[i] : 0) + delta[i];
-            if (dw[i] < m) {
-                m = -dw[i];
-            }
-        }
-        if (m) {
-            for (const i of dw.keys()) {
-                dw[i] += m;
-            }
-        }
-        return dw;
-    }
-
-    /**
      * @param {N[]} nodes    The HTML elements to be centered in a stack
      * @param {number[]} dx  The x offsets needed to center the elements
      */
@@ -282,54 +74,6 @@ export class CHTMLscriptbase<N, T, D> extends CHTMLWrapper<N, T, D> {
         for (let i = 0; i < dx.length; i++) {
             if (dx[i]) {
                 this.adaptor.setStyle(nodes[i], 'paddingLeft', this.em(dx[i]));
-            }
-        }
-    }
-
-    /**
-     * @return {number}   The offset for under and over
-     */
-    protected getDelta(noskew: boolean = false) {
-        const accent = this.node.attributes.get('accent');
-        const ddelta = (accent ? this.baseChild.coreMO().bbox.sk : 0);
-        return DELTA * this.baseCore.bbox.ic / 2 + ddelta;
-    }
-
-    /**
-     * Handle horizontal stretching of children to match greatest width
-     *  of all children
-     */
-    protected stretchChildren() {
-        let stretchy: CHTMLWrapper<N, T, D>[] = [];
-        //
-        //  Locate and count the stretchy children
-        //
-        for (const child of this.childNodes) {
-            if (child.canStretch(DIRECTION.Horizontal)) {
-                stretchy.push(child);
-            }
-        }
-        let count = stretchy.length;
-        let nodeCount = this.childNodes.length;
-        if (count && nodeCount > 1) {
-            let W = 0;
-            //
-            //  If all the children are stretchy, find the largest one,
-            //  otherwise, find the width of the non-stretchy children.
-            //
-            let all = (count > 1 && count === nodeCount);
-            for (const child of this.childNodes) {
-                const noStretch = (child.stretch.dir === DIRECTION.None);
-                if (all || noStretch) {
-                    const {w, rscale} = child.getBBox(noStretch);
-                    if (w * rscale > W) W = w * rscale;
-                }
-            }
-            //
-            //  Stretch the stretchable children
-            //
-            for (const child of stretchy) {
-                child.coreMO().getStretchedVariant([W / child.bbox.rscale]);
             }
         }
     }
